+++ bot/feeds_handler.go
@@ -15,7 +15,7 @@ import (
 func (ub *UndercastBot) feedsHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
 	userID := ub.extractUsername(update)
 	zapFields := []zap.Field{
-		zap.Int("chatID", update.Message.Chat.ID),
+		zap.Int64("chatID", update.Message.Chat.ID),
 		zap.String("messageText", update.Message.Text),
 		zap.String("userID", userID),
 	}
